@@ -1,4 +1,4 @@
-using System.Data.SqlClient;
+﻿using Microsoft.Data.SqlClient;
 
 namespace SS.Backend.DataAccess
 {
@@ -8,14 +8,12 @@ namespace SS.Backend.DataAccess
         ICustomSqlCommandBuilder Columns(IEnumerable<string> columns);
         ICustomSqlCommandBuilder Values(IEnumerable<string> columns);
         ICustomSqlCommandBuilder BeginUpdate(string tableName);
-        ICustomSqlCommandBuilder Select();
-        ICustomSqlCommandBuilder BeginSelect();
-        ICustomSqlCommandBuilder From(string tableName);
-        ICustomSqlCommandBuilder SelectOne(string column);
         ICustomSqlCommandBuilder Set(Dictionary<string, object> columnValues);
         ICustomSqlCommandBuilder Where(string whereClause);
+        ICustomSqlCommandBuilder BeginSelect();
         ICustomSqlCommandBuilder BeginSelectAll();
         ICustomSqlCommandBuilder SelectColumns(params string[] columns);
+        ICustomSqlCommandBuilder From(string tableName);
         ICustomSqlCommandBuilder BeginDelete(string tableName);
         ICustomSqlCommandBuilder Join(string joinTable, string fromColumn, string toColumn);
         ICustomSqlCommandBuilder AddParameters(Dictionary<string, object> parameters);
