@@ -1,6 +1,7 @@
 using SS.Backend.DataAccess;
 using SS.Backend.SharedNamespace;
-using System.Data.SqlClient;
+using Microsoft.Data.SqlClient;
+
 
 namespace SS.Backend.UserManagement
 {
@@ -8,12 +9,14 @@ namespace SS.Backend.UserManagement
     public class UserManagementRepository : IUserManagementRepository
     {
 
-        Credential removeMeLater = Credential.CreateSAUser();
-
-
         public async Task<Response> GeneralModifier(string whereClause, object whereClauseval, string fieldName, object newValue, string tableName)
         {
-            SealedSqlDAO SQLDao = new SealedSqlDAO(removeMeLater);
+            string configFilePath = "/Users/carinegordillo/config.txt";
+            ConfigService configService = new ConfigService(configFilePath);
+            SqlDAO sqldao = new SqlDAO(configService);
+
+
+
             Response response = new Response();
             var commandBuilder = new CustomSqlCommandBuilder();
             
@@ -21,13 +24,13 @@ namespace SS.Backend.UserManagement
             var columnValues = new Dictionary<string, object>{
                 { fieldName, newValue },{whereClause,whereClauseval}};
 
-            SqlCommand updateCommand = commandBuilder.BeginUpdate(tableName)
+            var updateCommand = commandBuilder.BeginUpdate(tableName)
                                             .Set(columnValues)
                                             .Where($"{whereClause} = @{whereClause}")
                                             .AddParameters(columnValues)
                                             .Build();
 
-            response = await SQLDao.SqlRowsAffected(updateCommand);
+            response = await sqldao.SqlRowsAffected(updateCommand);
 
             if (response.HasError == false){
                 response.ErrorMessage += "- General Modifier - command successful -";
@@ -42,17 +45,20 @@ namespace SS.Backend.UserManagement
         public async Task<Response> ReadUserTable(string tableName)
         {
 
-            SealedSqlDAO SQLDao = new SealedSqlDAO(removeMeLater);
+            string configFilePath = "/Users/carinegordillo/config.txt";
+            ConfigService configService = new ConfigService(configFilePath);
+            SqlDAO sqldao = new SqlDAO(configService);
+
             Response response = new Response();
             var commandBuilder = new CustomSqlCommandBuilder();
             
 
 
-            SqlCommand selectRequestsCommand = commandBuilder.BeginSelectAll()
+            var selectRequestsCommand = commandBuilder.BeginSelectAll()
                                             .From($"{tableName}")
                                             .Build();
 
-            response = await SQLDao.ReadSqlResult(selectRequestsCommand);
+            response = await sqldao.ReadSqlResult(selectRequestsCommand);
 
             if (response.HasError == false){
                 response.ErrorMessage += "- ReadUserTable- command successful -";
@@ -68,8 +74,12 @@ namespace SS.Backend.UserManagement
 
         public async Task<Response> createAccountRecoveryRequest(UserRequestModel userRequest, string tableName)
         {
-            SealedSqlDAO SQLDao = new SealedSqlDAO(removeMeLater);
+
+            string configFilePath = "/Users/carinegordillo/config.txt";
+            ConfigService configService = new ConfigService(configFilePath);
+            SqlDAO sqldao = new SqlDAO(configService);
             Response response = new Response();
+
             var commandBuilder = new CustomSqlCommandBuilder();
 
             var parameters = new Dictionary<string, object>
@@ -81,23 +91,93 @@ namespace SS.Backend.UserManagement
                             { "additionalInformation", userRequest.AdditionalInformation }
                         };
 
-            SqlCommand InsertRequestsCommand = commandBuilder.BeginInsert(tableName)
+            var InsertRequestsCommand = commandBuilder.BeginInsert(tableName)
                                                             .Columns(parameters.Keys)
                                                             .Values(parameters.Keys)
                                                             .AddParameters(parameters)
                                                             .Build();
 
-            response = await SQLDao.SqlRowsAffected(InsertRequestsCommand);
+            response = await sqldao.SqlRowsAffected(InsertRequestsCommand);
 
             if (response.HasError == false){
                 response.ErrorMessage += "- createAccountRecoveryRequest - command successful - ";
             }
             else{
-                 response.ErrorMessage += $"- createAccountRecoveryRequest - command : {InsertRequestsCommand} not successful - ";
+                 response.ErrorMessage += $"- createAccountRecoveryRequest - command : {InsertRequestsCommand.CommandText} not successful - ";
 
             }
             return response;
         }
+
+
+
+        public async Task<Response> sendRequest(string employeeName, string position)
+        {
+            
+
+            string configFilePath = "/Users/carinegordillo/config.txt";
+            ConfigService configService = new ConfigService(configFilePath);
+            SqlDAO sqldao = new SqlDAO(configService);
+            Response response = new Response();
+            var commandBuilder = new CustomSqlCommandBuilder();
+
+            var parameters = new Dictionary<string, object>
+                        {
+                            { "Name", employeeName },
+                            { "Position", position}
+                        };
+
+            var InsertRequestsCommand = commandBuilder.BeginInsert("dbo.EmployeesDummyTable")
+                                                            .Columns(parameters.Keys)
+                                                            .Values(parameters.Keys)
+                                                            .AddParameters(parameters)
+                                                            .Build();
+
+            response = await sqldao.SqlRowsAffected(InsertRequestsCommand);
+
+            if (response.HasError == false){
+                response.ErrorMessage += "- sendRequest - command successful - ";
+            }
+            else{
+                 response.ErrorMessage += $"- sendRequest - command : {InsertRequestsCommand.CommandText} not successful - ";
+
+            }
+            return response;
+        }
+
+        public async Task<Response> readTableWhere(string whereClause, object whereClauseval, string tableName)
+        {
+            string configFilePath = "/Users/carinegordillo/config.txt";
+            ConfigService configService = new ConfigService(configFilePath);
+            SqlDAO sqldao = new SqlDAO(configService);
+            Response response = new Response();
+            var commandBuilder = new CustomSqlCommandBuilder();
+
+            var parameters = new Dictionary<string, object>
+            {{ whereClause, whereClauseval }};
+
+            var command = commandBuilder.BeginSelectAll()
+                        .From(tableName)
+                        .Where($"{whereClause} = @{whereClause}")
+                        .AddParameters(parameters)
+                        .Build();
+            
+            response = await sqldao.ReadSqlResult(command);
+
+            if (response.HasError == false){
+                response.ErrorMessage += "- readTableWhere- command successful -";
+            }
+            else{
+                 response.ErrorMessage += $"- readTableWhere- {command.CommandText} -  command not successful -";
+
+            }
+            return response;
+
+
+
+        }
+
+
 
     }
 }
