@@ -4,18 +4,11 @@ namespace SS.Backend.DataAccess
 {
     public class ConfigService
     {
-        private readonly Dictionary<string, string> configValues;
+        private readonly string configValues;
 
         public ConfigService(string configFilePath)
         {
             configValues = LoadConfig(configFilePath);
-
-            // Print the loaded configuration values to the console for debugging
-            Console.WriteLine("Loaded Configuration Values:");
-            foreach (var kvp in configValues)
-            {
-                Console.WriteLine($"{kvp.Key}: {kvp.Value}");
-            }
         }
 
         /// <summary>
@@ -24,16 +17,7 @@ namespace SS.Backend.DataAccess
         /// <returns>The connection string</returns>
         public string GetConnectionString()
         {
-            Console.WriteLine("Loaded Configuration Values:");
-            foreach (var kvp in configValues)
-            {
-                Console.WriteLine($"{kvp.Key} = {kvp.Value}");
-            }
-
-            if (configValues.TryGetValue("ConnectionString", out var connectionString))
-            {
-                return connectionString;
-            }
+            return configValues;
 
             throw new InvalidOperationException("ConnectionString not found in the configuration file.");
         }
@@ -43,21 +27,15 @@ namespace SS.Backend.DataAccess
         /// </summary>
         /// <param name="filePath">The path of where the config file is on your system</param>
         /// <returns>Each line read in</returns>
-        private Dictionary<string, string> LoadConfig(string filePath)
+        private string LoadConfig(string filePath)
         {
-            var config = new Dictionary<string, string>();
-
+            string config = "";
             try
             {
-                Console.WriteLine($"Reading configuration from file: {filePath}");
-
-                // Read the first line as the connection string
                 var lines = File.ReadAllLines(filePath, Encoding.UTF8);
                 if (lines.Length > 0)
                 {
-                    var connectionString = lines[0].Trim();
-                    config["ConnectionString"] = connectionString;
-                    Console.WriteLine($"Loaded: ConnectionString = {connectionString}");
+                    config = lines[0];
                 }
             }
             catch (Exception ex)
@@ -67,6 +45,7 @@ namespace SS.Backend.DataAccess
             }
 
             return config;
+
         }
 
 
