@@ -1,20 +1,28 @@
+﻿using Microsoft.IdentityModel.Tokens;
 using SS.Backend.DataAccess;
+using SS.Backend.Services.LoggingService;
 using SS.Backend.SharedNamespace;
+using System.IdentityModel.Tokens.Jwt;
+using System.Security.Claims;
+using System.Text;
 
 namespace SS.Backend.Security
 {
-    public class SSAuthService : IAuthenticator, IAuthorizer
+    public class SSAuthService : IAuthenticator, IAuthorizer //add interfaces for the objects
     {
-        
         private readonly GenOTP genotp;
         private readonly Hashing hasher;
         private readonly SqlDAO sqldao;
+        private readonly Logger log;
+        private readonly string jwtSecret;
 
-        public SSAuthService(GenOTP genotp, Hashing hasher, SqlDAO sqldao)
+        public SSAuthService(GenOTP genotp, Hashing hasher, SqlDAO sqldao, Logger log, string jwtSecret)
         {
             this.genotp = genotp;
             this.hasher = hasher;
             this.sqldao = sqldao;
+            this.log = log;
+            this.jwtSecret = jwtSecret;
         }
 
         /// <summary>
@@ -33,14 +41,24 @@ namespace SS.Backend.Security
                 // check if user is registered, ie if user has a role
                 var selectCommand = builder
                     .BeginSelectAll()
-                    .From("UserRoles")
-                    .Where($"Username = '{user}'")
+                    .From("userProfile")
+                    .Where($"hashedUsername = '{user}'")
                     .Build();
-                result = await sqldao.ReadSqlResult(selectCommand).ConfigureAwait(false);
-                if (result.ValuesRead == null || result.ValuesRead.Count == 0)
+                result = await sqldao.ReadSqlResult(selectCommand);
+                if (result.ValuesRead == null || result.ValuesRead.Rows.Count == 0)
                 {
                     result.HasError = true;
                     result.ErrorMessage = "User does not exist.";
+                    LogEntry entry = new()
+                    {
+                        timestamp = DateTime.UtcNow,
+                        level = "Error",
+                        username = user,
+                        category = "Data Store",
+                        description = "Unregistered user tried to authenticate."
+                    };
+                    await log.SaveData(entry);
+
                     return (null, result);
                 }
 
@@ -57,8 +75,8 @@ namespace SS.Backend.Security
                     .From("OTP")
                     .Where($"Username = '{user}'")
                     .Build();
-                result = await sqldao.ReadSqlResult(otpRead).ConfigureAwait(false);
-                if (result.ValuesRead == null || result.ValuesRead.Count == 0)
+                result = await sqldao.ReadSqlResult(otpRead);
+                if (result.ValuesRead == null || result.ValuesRead.Rows.Count == 0)
                 {
                     // create and execute the sql command to insert the username, otp, salt, and timestamp to the DB
                     var parameters = new Dictionary<string, object>
@@ -74,9 +92,19 @@ namespace SS.Backend.Security
                         .Values(parameters.Keys)
                         .AddParameters(parameters)
                         .Build();
-                    result = await sqldao.SqlRowsAffected(insertCommand).ConfigureAwait(false);
+                    result = await sqldao.SqlRowsAffected(insertCommand);
 
                     // user and otp is returned so it can be used by the manager to send the otp to the user
+                    LogEntry entry = new()
+                    {
+                        timestamp = DateTime.UtcNow,
+                        level = "Info",
+                        username = user,
+                        category = "Data Store",
+                        description = "Successfully sent OTP to manager."
+                    };
+                    await log.SaveData(entry);
+
                     return (otp, result);
                 }
                 else
@@ -96,9 +124,19 @@ namespace SS.Backend.Security
                         .Build();
 
                     //update the otp and salt in table for that user
-                    result = await sqldao.SqlRowsAffected(updateCommand).ConfigureAwait(false);
+                    result = await sqldao.SqlRowsAffected(updateCommand);
 
                     // user and otp is returned so it can be used by the manager to send the otp to the user
+                    LogEntry entry = new()
+                    {
+                        timestamp = DateTime.UtcNow,
+                        level = "Info",
+                        username = user,
+                        category = "Data Store",
+                        description = "Successfully sent OTP to manager."
+                    };
+                    await log.SaveData(entry);
+
                     return (otp, result);
                 }
             }
@@ -148,10 +186,10 @@ namespace SS.Backend.Security
                     .From("OTP")
                     .Where($"Username = '{hashedUsername}'")
                     .Build();
-                result = await sqldao.ReadSqlResult(selectCommand).ConfigureAwait(false);
-                string dbOTP = (string)result.ValuesRead[0][1];
-                string dbSalt = (string)result.ValuesRead[0][2];
-                DateTime timestamp = (DateTime)result.ValuesRead[0][3];
+                result = await sqldao.ReadSqlResult(selectCommand);
+                string dbOTP = (string)result.ValuesRead.Rows[0]["OTP"];
+                string dbSalt = (string)result.ValuesRead.Rows[0]["Salt"];
+                DateTime timestamp = (DateTime)result.ValuesRead.Rows[0]["Timestamp"];
                 TimeSpan timeElapsed = DateTime.UtcNow - timestamp;
 
                 // compare the otp stored in DB with user inputted otp
@@ -162,6 +200,16 @@ namespace SS.Backend.Security
                     {
                         result.HasError = true;
                         result.ErrorMessage = "OTP has expired.";
+                        LogEntry entry = new()
+                        {
+                            timestamp = DateTime.UtcNow,
+                            level = "Error",
+                            username = hashedUsername,
+                            category = "Data Store",
+                            description = "User tried to authenticate with an expired OTP."
+                        };
+                        await log.SaveData(entry);
+
                         return (null, result);
                     }
                     else
@@ -169,13 +217,13 @@ namespace SS.Backend.Security
                         // they match and not expired, so get the roles from the DB for that user
                         var readRoles = builder
                             .BeginSelectAll()
-                            .From("UserRoles")
-                            .Where($"Username = '{hashedUsername}'")
+                            .From("userProfile")
+                            .Where($"hashedUsername = '{hashedUsername}'")
                             .Build();
-                        result = await sqldao.ReadSqlResult(readRoles).ConfigureAwait(false);
-                        if (result.ValuesRead.Count > 0)
+                        result = await sqldao.ReadSqlResult(readRoles);
+                        if (result.ValuesRead.Rows.Count > 0)
                         {
-                            string role = (string)result.ValuesRead[0][1];
+                            string role = result.ValuesRead.Rows[0]["appRole"].ToString();
 
                             // populate the principal
                             SSPrincipal principal = new SSPrincipal();
@@ -184,12 +232,32 @@ namespace SS.Backend.Security
 
                             result.HasError = false;
 
+                            LogEntry entry = new()
+                            {
+                                timestamp = DateTime.UtcNow,
+                                level = "Info",
+                                username = hashedUsername,
+                                category = "Data Store",
+                                description = "Successful authentication."
+                            };
+                            await log.SaveData(entry);
+
                             return (principal, result);
                         }
                         else
                         {
                             result.HasError = true;
                             result.ErrorMessage = "No roles found for the user.";
+                            LogEntry entry = new()
+                            {
+                                timestamp = DateTime.UtcNow,
+                                level = "Error",
+                                username = hashedUsername,
+                                category = "Data Store",
+                                description = "Failure to authenticate."
+                            };
+                            await log.SaveData(entry);
+
                             return (null, result);
                         }
                     }
@@ -199,12 +267,32 @@ namespace SS.Backend.Security
                 {
                     result.HasError = true;
                     result.ErrorMessage = "Failed to authenticate.";
+                    LogEntry entry = new()
+                    {
+                        timestamp = DateTime.UtcNow,
+                        level = "Error",
+                        username = hashedUsername,
+                        category = "Data Store",
+                        description = "Failure to authenticate."
+                    };
+                    await log.SaveData(entry);
+
                     return (null, result);
                 }
             }
             catch (Exception ex)
             {
                 result.ErrorMessage = ex.Message;
+                LogEntry entry = new()
+                {
+                    timestamp = DateTime.UtcNow,
+                    level = "Error",
+                    username = hashedUsername,
+                    category = "Data Store",
+                    description = "Failure to authenticate."
+                };
+                await log.SaveData(entry);
+
                 return (null, result);
             }
 
@@ -218,6 +306,12 @@ namespace SS.Backend.Security
         /// <returns>A boolean denoting if the user is authorized</returns>
         public async Task<bool> IsAuthorize(SSPrincipal currentPrincipal, IDictionary<string, string> requiredClaims)
         {
+            if (currentPrincipal?.Claims == null)
+            {
+                // If user claims are null, authorization fails
+                return false;
+            }
+
             foreach (var claim in requiredClaims)
             {
                 if (!currentPrincipal.Claims.Contains(claim))
@@ -227,6 +321,45 @@ namespace SS.Backend.Security
             }
             return true;
 
+        }
+
+        public SSPrincipal MapToSSPrincipal(ClaimsPrincipal claimsPrincipal)
+        {
+            if (claimsPrincipal == null)
+            {
+                return null;
+            }
+
+            var ssPrincipal = new SSPrincipal
+            {
+                UserIdentity = claimsPrincipal.Identity?.Name,
+                Claims = new Dictionary<string, string>()
+            };
+
+            foreach (var claim in claimsPrincipal.Claims)
+            {
+                ssPrincipal.Claims.Add(claim.Type, claim.Value);
+            }
+
+            return ssPrincipal;
+        }
+
+        private string GenerateJwtToken(string username, string role)
+        {
+            var tokenHandler = new JwtSecurityTokenHandler();
+            var key = Encoding.ASCII.GetBytes(jwtSecret);
+            var tokenDescriptor = new SecurityTokenDescriptor
+            {
+                Subject = new ClaimsIdentity(new[]
+                {
+                    new Claim(ClaimTypes.Name, username),
+                    new Claim(ClaimTypes.Role, role)
+                }),
+                Expires = DateTime.UtcNow.AddDays(7),
+                SigningCredentials = new SigningCredentials(new SymmetricSecurityKey(key), SecurityAlgorithms.HmacSha256Signature)
+            };
+            var token = tokenHandler.CreateToken(tokenDescriptor);
+            return tokenHandler.WriteToken(token);
         }
     }
 }
