@@ -1,9 +1,6 @@
 using SS.Backend.DataAccess;
 using SS.Backend.SharedNamespace;
-using System.Collections.Generic;
-using System.Threading.Tasks;
 using System.Reflection;
-using SS.Backend.Services.LoggingService;
 using System.Text.RegularExpressions;
 
 
@@ -56,32 +53,34 @@ namespace SS.Backend.Services.AccountCreationService
                         }
                         break;
                     case "companyName":
-                        if (userInfo.role == 2 || userInfo.role == 3){
-                                if(!IsValidCompanyName(value as string))
+                        if (userInfo.role == 2 || userInfo.role == 3)
+                        {
+                            if (!IsValidCompanyName(value as string))
                             {
                                 errorMsg += $"Invalid {prop.Name.ToLower()}; ";
                             }
                             break;
                         }
                         break;
-                        
+
                     case "address":
-                        if (userInfo.role == 2 || userInfo.role == 3){
-                            if(!IsValidAddress(value as string))
+                        if (userInfo.role == 2 || userInfo.role == 3)
+                        {
+                            if (!IsValidAddress(value as string))
                             {
                                 errorMsg += $"Invalid {prop.Name.ToLower()}; ";
                             }
                             break;
                         }
                         break;
-                    // case "openingHours":
-                    // case "closingHours":
-                    // case "daysOpen": //check if these need their own function
-                    //     if (CheckNullWhiteSpace(value as string))
-                    //     {
-                    //         errorMsg += $"Invalid {prop.Name.ToLower()}; ";
-                    //     }
-                    //     break;
+                        // case "openingHours":
+                        // case "closingHours":
+                        // case "daysOpen": //check if these need their own function
+                        //     if (CheckNullWhiteSpace(value as string))
+                        //     {
+                        //         errorMsg += $"Invalid {prop.Name.ToLower()}; ";
+                        //     }
+                        //     break;
                 }
             }
             return string.IsNullOrEmpty(errorMsg) ? "Pass" : errorMsg;
@@ -117,7 +116,7 @@ namespace SS.Backend.Services.AccountCreationService
         }
         private bool IsValidCompanyName(string name)
         {
-            return 
+            return
                 name.Length >= 1 &&
                 name.Length <= 60;
         }
@@ -130,9 +129,11 @@ namespace SS.Backend.Services.AccountCreationService
         //this method takes builds a dictionary with several sql commands to insert all at once 
         public async Task<Response> InsertIntoMultipleTables(Dictionary<string, Dictionary<string, object>> tableData)
         {
+
             
             // SealedSqlDAO SQLDao = new SealedSqlDAO(temp);
             SqlDAO SQLDao = new SqlDAO(configService);
+
             var builder = new CustomSqlCommandBuilder();
             Response tablesresponse = new Response();
 
@@ -141,13 +142,13 @@ namespace SS.Backend.Services.AccountCreationService
             {
                 string tableName = tableEntry.Key;
                 Dictionary<string, object> parameters = tableEntry.Value;
-                var insertCommand =  builder.BeginInsert(tableName)
+                var insertCommand = builder.BeginInsert(tableName)
                     .Columns(parameters.Keys)
                     .Values(parameters.Keys)
                     .AddParameters(parameters)
                     .Build();
 
-                tablesresponse = await SQLDao.SqlRowsAffected(insertCommand);
+                //tablesresponse = await SQLDao.SqlRowsAffected(insertCommand);
                 if (tablesresponse.HasError)
                 {
                     tablesresponse.ErrorMessage += $"{tableName}: error inserting data; ";
@@ -162,8 +163,10 @@ namespace SS.Backend.Services.AccountCreationService
         {
             Response response = new Response();
 
+
             SqlDAO SQLDao = new SqlDAO(configService);
             Logger logger = new Logger(new SqlLogTarget(new SqlDAO(configService)));
+
 
 
             string validationMessage = CheckUserInfoValidity(userInfo);
@@ -173,8 +176,9 @@ namespace SS.Backend.Services.AccountCreationService
                 response.ErrorMessage = "Invalid User Info entry: " + validationMessage;
                 return response;
             }
-       
+
             //generating sql command 
+
             UserPepper userPepper = new UserPepper();
             AccountCreation accountcreation = new AccountCreation();
             Hashing hashing = new Hashing();
@@ -227,6 +231,7 @@ namespace SS.Backend.Services.AccountCreationService
 
 
             response  = await InsertIntoMultipleTables(tableData);
+
             if (response.HasError == false)
             {
                 LogEntry entry = new LogEntry()
@@ -238,9 +243,10 @@ namespace SS.Backend.Services.AccountCreationService
                     category = "Data Store",
                     description = "Successful account creation"
                 };
-                await logger.SaveData(entry);
+                //await logger.SaveData(entry);
             }
-            else{
+            else
+            {
                 LogEntry entry = new LogEntry()
 
                 {
@@ -250,10 +256,10 @@ namespace SS.Backend.Services.AccountCreationService
                     category = "Data Store",
                     description = "Error inserting user in data store."
                 };
-                await logger.SaveData(entry);
+                // await logger.SaveData(entry);
             }
-            return response;            
-         
+            return response;
+
         }
 
         public async Task<Response> ReadUserTable(string tableName)
