@@ -10,8 +10,8 @@ namespace SS.Backend.Services.DeletingService
     ///
     public class AccountDeletion : IAccountDeletion
     {
-        // Temporary User Variable
-        Credential temp = Credential.CreateSAUser();
+
+        ConfigService configService = new ConfigService("C:/Users/brand/Documents/GitHub/SpaceSurfer/SourceCode/SS.Backend/config.local.txt");
 
         /// <summary>
         ///     DeleteAccount deletes the account by username
@@ -21,93 +21,110 @@ namespace SS.Backend.Services.DeletingService
         public async Task<Response> DeleteAccount(string username)
         {
             // initializes a new instance of the Response
-            Response overallResponse = new Response();
-
-            // initializes a new instance of SqlDAO
-            SealedSqlDAO SQLDao = new SealedSqlDAO(temp);
-
-            // initializes a new instance of the logger
-            Logger logger = new Logger(new SqlLogTarget(new SqlDAO(temp)));
-
-            // initializes a new instance of the Custom Command Builder
-            CustomSqlCommandBuilder commandBuilder = new CustomSqlCommandBuilder();
-
-            // initializes a new instance of Database Helper
-            DatabaseHelper dbHelper = new DatabaseHelper();
-
-            // Sets the response error to false
-            overallResponse.HasError = false;
-
-            // Sets the tables names from the Database Helper Response
-            var tables = dbHelper.RetrieveTableNames();
-
-            // Sets the value to the username
-            var value = new Dictionary<string, object>
-            {
-                { "Username", username}
-            };
+            Response result = new Response();
 
             try
             {
-                // Iterates through every table that deals with the value being deleted
-                for (int i = tables.Result.ValuesRead.Count - 1; i >= 0; i--)
+
+                // initializes a new instance of SqlDAO
+                SqlDAO SQLDao = new SqlDAO(configService);
+
+                // initializes a new instance of the logger
+                Logger logger = new Logger(new SqlLogTarget(new SqlDAO(configService)));
+
+                // initializes a new instance of the Custom Command Builder
+                var commandBuilder = new CustomSqlCommandBuilder();
+
+                var command = commandBuilder.BeginSelectAll()
+                        .From("dbo.userAccount ")
+                        .Where($"Username = '{username}'")
+                        .Build();
+                // initializes a new instance of Database Helper
+                //DatabaseHelper dbHelper = new DatabaseHelper();
+
+                // Sets the tables names from the Database Helper Response
+                result = await SQLDao.ReadSqlResult(command);
+
+                // Sets the value to the username
+                //var value = new Dictionary<string, object>
+                //{
+                //    { "Username", username}
+                //};
+
+
+                if (result.ValuesRead != null || result.ValuesRead.Rows.Count != null)
                 {
-                    // Initializing Table Name 
-                    var table = tables.Result.ValuesRead[i][2];
 
                     // Delete Query Command built [DELETE FROM "Users" WHERE Username = @username]
-                    var deleteCommand = commandBuilder.BeginDelete("dbo." + table).Where("Username = @Username").AddParameters(value).Build();
+                    var deleteCommand = commandBuilder.BeginDelete("dbo.userAccount").Where($"Username = '{username}'").Build();
 
                     // Sets the reponse from the executed command
-                    var response = await SQLDao.SqlRowsAffected(deleteCommand);
+                    result = await SQLDao.SqlRowsAffected(deleteCommand);
 
-                    // Appends the responses into one response
-                    overallResponse.RowsAffected += response.RowsAffected;
-                    overallResponse.HasError |= response.HasError;
-
+                    //// Appends the responses into one response
+                    //overallResponse.RowsAffected += response.RowsAffected;
+                    //overallResponse.HasError |= response.HasError;
+                }
+                else
+                {
+                    // Handle the case where tables.Task is not completed or its result is null
+                    throw new InvalidOperationException("Failed to retrieve table.");
                 }
 
+                // await tables;
+                //// Check if the task completed successfully and its result is not null
+                //// if (tables != null && tables.Result != null && tables.Result.ValuesRead != null)
+                //// {
+                //    // Iterates through every table that deals with the value being deleted
+                //    for (int i = tables.Result.ValuesRead.Count - 1; i >= 0; i--)
+                //    {
+                //        
+
+                //    }
+                //}
+
+                if (result.HasError == false)
+                {
+
+                    // Successful Deletion
+                    LogEntry entry = new LogEntry()
+
+                    {
+                        timestamp = DateTime.UtcNow,
+                        level = "Info",
+                        username = username,
+                        category = "Data Store",
+                        description = "Successful Deletion"
+                    };
+
+                    await logger.SaveData(entry);
+                }
+                else
+                {
+                    //Unsuccessful Deletion
+                    LogEntry errorEntry = new LogEntry()
+                    {
+                        timestamp = DateTime.UtcNow,
+                        level = "Error",
+                        username = username,
+                        category = "Data Store",
+                        description = "Unsuccessful Deletion"
+                    };
+
+                    await logger.SaveData(errorEntry);
+                }
             }
             catch (Exception ex)
             {
                 // If an error occurs error
-                overallResponse.HasError = true;
-                overallResponse.ErrorMessage = ex.Message;
+                result.HasError = true;
+                result.ErrorMessage = ex.Message;
             }
 
             // Logs entry based on the overallResponse error
-            if (overallResponse.HasError == false)
-            {
 
-                // Successful Deletion
-                LogEntry entry = new LogEntry()
 
-                {
-                    timestamp = DateTime.UtcNow,
-                    level = "Info",
-                    username = username,
-                    category = "Data Store",
-                    description = "Successful Deletion"
-                };
-
-                await logger.SaveData(entry);
-            }
-            else
-            {
-                //Unsuccessful Deletion
-                LogEntry errorEntry = new LogEntry()
-                {
-                    timestamp = DateTime.UtcNow,
-                    level = "Error",
-                    username = username,
-                    category = "Data Store",
-                    description = "Unsuccessful Deletion"
-                };
-
-                await logger.SaveData(errorEntry);
-            }
-
-            return overallResponse;
+            return result;
         }
 
     }
