@@ -4,9 +4,10 @@
     {
         public DateTime timestamp { get; set; }
         public string level { get; set; }
-        public string username { get; set; }
+        public string? username { get; set; }
         public string category { get; set; }
         public string description { get; set; }
+
 
         public int getDescriptionLength
         {
