@@ -1,3 +1,4 @@
+
 using SS.Backend.SharedNamespace;
 
 
@@ -5,6 +6,15 @@ namespace SS.Backend.UserManagement
 {
     public class AccountRecoveryNoInj : IAccountRecovery
     {
+
+
+        /*
+         * This method creates a recovery request for a user.
+         * automatically sends the request status to pending
+         * @param userHash The user's hash.
+         * @param additionalInfo Additional information to be added to the request.
+         * @return Response object.
+         */
 
 
         public async Task<Response> createRecoveryRequest(string userHash, string additionalInfo = "")
@@ -39,6 +49,11 @@ namespace SS.Backend.UserManagement
         }
         
 
+        /* 
+         * This method sends a recovery request to the userManagementRepository, updates activeAccount table to mark account as pending
+         * @param userHash - the hashed username of the user
+         * @return Response - the response object
+         */
 
         public async Task<Response> sendRecoveryRequest(string userHash)
         {
@@ -49,15 +64,21 @@ namespace SS.Backend.UserManagement
 
             if (response.HasError == false)
             {
-                response.ErrorMessage = "- Recovery request initiated. -";
+                response.ErrorMessage = "- Recovery request sent. -";
             }
             else
             {
-                response.ErrorMessage = "- Failed to initiate recovery request.- ";
+                response.ErrorMessage = "- Failed to send recovery request.- ";
             }
 
             return response;
         }
+
+        /*
+        * This method is used to recover a user account, and update the status of the user request in the userRequests table to accepted or denied
+        * @param userHash - the hashed username of the user
+        * @param adminDecision - the decision of the admin to accept or deny the request
+        */
 
         public async Task<Response> RecoverAccount(string userHash, bool adminDecision)
         {
@@ -76,6 +97,12 @@ namespace SS.Backend.UserManagement
             return response;
             
         }
+
+        /* 
+        *
+        * This method is used to read all the userRequests table from the database
+        * @return Response - the response object
+        */
 
         public async Task<Response> ReadUserRequests(){
 
@@ -96,6 +123,65 @@ namespace SS.Backend.UserManagement
 
             return response;
         }
+
+         public async Task<Response> ReadUserPendingRequests(){
+
+            IUserManagementRepository userManagementRepository = new UserManagementRepository();
+            
+            Response response = new Response();
+            
+            response = await userManagementRepository.readTableWhere("status", "Pending", "dbo.userRequests");
+
+            if (response.HasError == false)
+            {
+                response.ErrorMessage += "- ReadUserPendingRequests successful. -";
+            }
+            else
+            {
+                response.ErrorMessage += "- ReadUserPendingRequests Failed - ";
+            }
+
+            return response;
+        }
+        
+        
+
+        public async Task<Response> ReadDummyTable(){
+
+            IUserManagementRepository userManagementRepository = new UserManagementRepository();
+            Response response = new Response();
+            
+            response = await userManagementRepository.ReadUserTable("dbo.EmployeesDummyTable");
+
+            if (response.HasError == false)
+            {
+                response.ErrorMessage += "- ReadDummyTable successful. -";
+            }
+            else
+            {
+                response.ErrorMessage += "- ReadDummyTable Failed - ";
+            }
+
+            return response;
+        }
+
+        public async Task<Response> sendDummyRequest(string name, string position)
+        {
+
+            IUserManagementRepository userManagementRepository = new UserManagementRepository();
+
+            Response response = new Response();
+            response = await userManagementRepository.sendRequest(name, position);
+
+            return response;
+
+        }
+
+        
+
+        
+
+
 
         
 
